+++ src/main/java/com/careydevelopment/twitterautomation/util/AnchorTextDataParser.java
@@ -1,5 +1,7 @@
 package com.careydevelopment.twitterautomation.util;
 
+import java.math.BigInteger;
+
 import com.careydevelopment.twitterautomation.jpa.entity.AnchorTextData;
 
 public class AnchorTextDataParser {
@@ -13,9 +15,9 @@ public class AnchorTextDataParser {
 				data.setCitationFlow(new Integer(parts[5]));
 				data.setTrustFlow(new Integer(parts[6]));
 				data.setReferringDomains(new Integer(parts[1]));
-				data.setTotalLinks(new Integer(parts[2]));
-				data.setDeletedLinks(new Integer(parts[3]));
-				data.setNofollowLinks(new Integer(parts[4]));
+				data.setTotalLinks(new BigInteger(parts[2]));
+				data.setDeletedLinks(new BigInteger(parts[3]));
+				data.setNofollowLinks(new BigInteger(parts[4]));
 				data.setAnchorText(parts[0]);
 			} else {
 				throw new RuntimeException("Couldn't parse the data row for anchor text!");
