+++ src/main/java/com/careydevelopment/twitterautomation/util/IndexItemInfoRowParser.java
@@ -1,5 +1,7 @@
 package com.careydevelopment.twitterautomation.util;
 
+import java.math.BigInteger;
+
 import com.careydevelopment.twitterautomation.jpa.entity.IndexItemInfoRow;
 
 public class IndexItemInfoRowParser {
@@ -14,18 +16,18 @@ public class IndexItemInfoRowParser {
 				iir.setItem(parts[1]);
 				iir.setResultCode(parts[2]);
 				iir.setStatus(parts[3]);
-				iir.setExtBacklinks(new Integer(parts[4]));
-				iir.setRefDomains(new Integer(parts[5]));
+				iir.setExtBacklinks(new BigInteger(parts[4]));
+				iir.setRefDomains(new BigInteger(parts[5]));
 				iir.setItemType(new Integer(parts[8]));
-				iir.setRefIps(new Integer(parts[12]));
-				iir.setRefDomainsEdu(new Integer(parts[14]));
-				iir.setExtBacklinksEdu(new Integer(parts[15]));
-				iir.setRefDomainsGov(new Integer(parts[16]));
-				iir.setExtBacklinksGov(new Integer(parts[17]));
-				iir.setRefDomainsEduExact(new Integer(parts[18]));
-				iir.setExtBacklinksEduExact(new Integer(parts[19]));
-				iir.setRefDomainsGovExact(new Integer(parts[20]));
-				iir.setExtBacklinksGovExact(new Integer(parts[21]));
+				iir.setRefIps(new BigInteger(parts[12]));
+				iir.setRefDomainsEdu(new BigInteger(parts[14]));
+				iir.setExtBacklinksEdu(new BigInteger(parts[15]));
+				iir.setRefDomainsGov(new BigInteger(parts[16]));
+				iir.setExtBacklinksGov(new BigInteger(parts[17]));
+				iir.setRefDomainsEduExact(new BigInteger(parts[18]));
+				iir.setExtBacklinksEduExact(new BigInteger(parts[19]));
+				iir.setRefDomainsGovExact(new BigInteger(parts[20]));
+				iir.setExtBacklinksGovExact(new BigInteger(parts[21]));
 				iir.setTitle(parts[31]);
 				iir.setCitationFlow(new Integer(parts[33]));
 				iir.setTrustFlow(new Integer(parts[34]));
